requirements.txt
qrcode==7.4.2
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
